view-2d: allow plugins to derive scale and translation dynamically
This way, crossfade can dynamically recalculate the scale and
translation when the view geometry changes, instead of waiting until the
next frame starts. Otherwise the bounding box may get very big (when the
view changes size mid-animation), and depending on the circumstances we
may try to generate a snapshot of this large view. This in turn leads to
a very large buffer allocation, which may fail and lead to crashes.