window-manager: Ignore unfullscreen requests if already windowed (#2764)
This fixes dragging gtk4 apps with grid and/or wobbly disabled. The gtk4
toolkit sends a request for unfullscreen even though it is windowed, and
this caused wayfire to set the geometry when dragging by titlebar, which
made the window jump to an offset when grabbing the titlebar to move it.
This was fixed before but broken by 60231d5d.